# public/openapi.yaml
openapi: 3.0.3
info:
  title: HookStep / Talent Platform API
  description: REST API for the hiring platform — jobs, applications, profile, and company dashboard.
  version: 1.0.0
servers:
  - url: /
    description: Current origin (relative)
...
paths:
...
  /api/upload/resume:
    post:
      tags: [Profile]
      summary: Upload resume PDF (replaces previous)
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF only, max 5MB
      responses:
        '200':
          description: New resume URL stored; previous deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  resumeUrl: { type: string, format: uri }
        '400':
          description: Missing file, not PDF, or too large
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ----- Utility -----
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie (sign in via the app)
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
      headers: {}
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...